day-1 commit

+++ transfer-service-v1/src/main/java/com/example/service/TransferService.java
@@ -1,5 +1,9 @@
 package com.example.service;
 
+import java.math.BigDecimal;
+
+// ISP — small, focused interface with single method
+// SRP — only defines the transfer operation contract
 public interface TransferService {
-    void transfer(double amount, String fromAccount, String toAccount);
+    void transfer(BigDecimal amount, String fromAccount, String toAccount);
 }
