+++ transfer-service-v2/src/main/java/com/example/service/TransferService.java
@@ -1,5 +1,8 @@
 package com.example.service;
 
+import java.math.BigDecimal;
+
+// ISP — small, focused interface with single method
 public interface TransferService {
-    void transfer(double amount, String fromAccount, String toAccount);
+    void transfer(BigDecimal amount, String fromAccount, String toAccount);
 }
